+++ Source/AccelByteUe4Sdk/Private/Api/AccelByteLobbyApi.cpp
@@ -16,6 +16,7 @@
 #include "Core/FUnrealWebSocketFactory.h"
 #include "Core/IAccelByteTokenGenerator.h"
 #include "Core/AccelByteError.h"
+#include "Core/AccelByteMessageParser.h"
 
 DEFINE_LOG_CATEGORY(LogAccelByteLobby);
 
@@ -686,8 +687,9 @@ void Lobby::SetPartySizeLimit(const FString& PartyId, const int32 Limit, const F
 FString Lobby::SendSetPresenceStatus(const EAvailability Availability, const FString& Activity)
 {
 	FReport::Log(FString(__FUNCTION__));
+	const FString EscapedActivity = MessageParser::EscapeString(Activity);
 	SEND_RAW_REQUEST_CACHED_RESPONSE_RETURNED(SetUserPresence, Presence
-		, FString::Printf(TEXT("availability: %s\nactivity: %s\n"), *FAccelByteUtilities::GetUEnumValueAsString(Availability).ToLower(), *Activity))
+		, FString::Printf(TEXT("availability: %s\nactivity: %s\n"), *FAccelByteUtilities::GetUEnumValueAsString(Availability).ToLower(), *EscapedActivity))
 }
 
 FString Lobby::SendGetOnlineUsersRequest()
@@ -1538,113 +1540,94 @@ void Lobby::CreateWebSocket(const FString& Token)
 	WebSocket->OnConnectionClosed().AddRaw(this, &Lobby::OnClosed);
 }
 
-FString Lobby::LobbyMessageToJson(FString Message)
+FString Lobby::LobbyMessageToJson(const FString& Message)
 {
-	FString Json = TEXT("{");
-	TArray<FString> Out;
-	Message.ParseIntoArray(Out, TEXT("\n"), true);
-	for (int i = 0; i < Out.Num(); i++)
+	bool bFirst = true;
+	FString JsonString = TEXT("{");
+	TArray<FString> Lines;
+	Message.ParseIntoArray(Lines, TEXT("\n"), true);
+	for (const auto& Line : Lines)
 	{
-		FString CurrentLine = Out[i];
-
-		FString Key;
+		FString Name;
 		FString Value;
-
-		CurrentLine.Split(": ", &Key, &Value);
-		Json += FString::Printf(TEXT("\"%s\":"), *Key);
-		Value.TrimStartAndEndInline();
-
-		if (Value.StartsWith("["))
+		if (bFirst)
 		{
-			// if starts with "[{" then it's an array of jsonObject field, pass the value since expected to be valid json field.
-			// "[]" is also a valid empty array json field.
-			if (Value.Equals("[]") || Value.StartsWith("[{")) 
-			{
-				Json += Value;
-			}
-			else
-			{
-				bool Quote = false;
-				bool ElementStart = false;
-				FString Element;
-				Json += "[";
-				for (int j = 1; j < Value.Len() - 1; j++)
-				{
-					if (!ElementStart)
-					{
-						if (Value[j] == ' ') 
-						{
-							continue;
-						}
-
-						ElementStart = true;
-						Element.AppendChar('"');
-						if (Value[j] == '"')
-						{
-							Quote = true;
-							continue;
-						}
-					}
-
-					if (!Quote && Value[j] == ',')
-					{
-						ElementStart = false;
-						Element.TrimEndInline();
-						Json += Element;
-
-						if (!Element.EndsWith("\"") // if element doesn't end with (")
-							|| (Element.Equals("\"") && Element.Len() == 1) // if element only (")
-							|| Element.Equals(",\"")) // if element only (,")
-						{
-							Json.AppendChar('"');
-						}
-						Element = "";
-
-						if (j == Value.Len() - 2) 
-						{
-							break;
-						}
-					}
-
-					Element.AppendChar(Value[j]);
-
-					if (Quote && Value[j] == '\\')
-					{
-						Element.AppendChar(Value[++j]);
-					}
-					else if (Value[j] == '"')
-					{
-						Quote = false;
-					}
-				}
-
-				if (!Element.IsEmpty())
-				{
-					Json += Element;
-					if (!Element.EndsWith("\""))
-					{
-						Json.AppendChar('"');
-					}
-				}
-				Json += "]";
-			}
-		}
-		else if (Value.StartsWith("{"))
-		{
-			Json += Value;
+			bFirst = false;
 		}
 		else
 		{
-			Json += FString::Printf(TEXT("\"%s\""), *Value);
+			JsonString.Append(",");
 		}
+		Line.Split(": ", &Name, &Value);
+		JsonString.Appendf(TEXT("\"%s\":"), *Name);
 
-		if (i < Out.Num() - 1)
+		Value.TrimStartAndEndInline();
+
+		const TCHAR* Cursor = GetData(Value);
+
+		if (Cursor == nullptr)
 		{
-			Json += ",";
+			JsonString.Append("null");
+			continue;
+		}
+		
+		// make sure it's null terminated
+		checkf(*(Cursor + Value.Len()) == 0, TEXT("Invalid value: '%s' length: %d"), *Value, Value.Len());
+
+		// Array
+		if (*Cursor == '[')
+		{
+			++Cursor;
+			// skip spaces
+			while (*Cursor && *Cursor == ' ') ++Cursor;
+			bool bWasArrayParsed;
+			FString JsonArrayString;
+			// array of JSON object
+			if (*Cursor == '{')
+			{
+				bWasArrayParsed = MessageParser::ParseArrayOfObject(Cursor, JsonArrayString);
+			}
+			// array of string
+			else
+			{
+				bWasArrayParsed = MessageParser::ParseArrayOfString(Cursor, JsonArrayString);
+			}
+			
+			if (bWasArrayParsed)
+			{
+				JsonString.Append(JsonArrayString);
+			}
+			else
+			{
+				// if the array was not parsed, set to empty array
+				JsonString.Append("[]");
+				UE_LOG(LogAccelByte, Warning, TEXT("[LobbyMessageToJson] Invalid array for field '%s', set to empty array"), *Name);
+			}
+		}
+		// JSON
+		else if (*Cursor == '{')
+		{
+			FString ObjectString;
+			// only append valid object
+			if (MessageParser::ParseObject(Cursor, ObjectString))
+			{
+				JsonString.Append(ObjectString);
+			}
+			else
+			{
+				JsonString.Append("{}");
+				UE_LOG(LogAccelByte, Warning, TEXT("[LobbyMessageToJson] Invalid object for field '%s', set to empty object"), *Name);
+			}
+		}
+		// everything else
+		else
+		{
+			MessageParser::ParseString(Cursor, JsonString);
 		}
 	}
-	Json += TEXT("}");
-	return Json;
+
+	JsonString += TEXT("}");
+	return JsonString;
 }
 
 /**
@@ -2212,7 +2195,6 @@ void Lobby::FetchLobbyErrorMessages()
 {
 	FString Url = FString::Printf(TEXT("%s/lobby/v1/messages"), *SettingsRef.BaseUrl);
 
-	bool bFetchErrorMessageDone = false;
 	FHttpClient HttpClient(CredentialsRef, SettingsRef, HttpRef);
 	HttpClient.Request("GET", Url, {}, "", {}, THandler<TArray<FLobbyMessages>>::CreateLambda([&](const TArray<FLobbyMessages>& Result)
 	{
@@ -2221,12 +2203,10 @@ void Lobby::FetchLobbyErrorMessages()
 			LobbyErrorMessages.Add(LobbyMessages.Code, LobbyMessages.CodeName);
 		}
 
-		bFetchErrorMessageDone = true;
 		UE_LOG(LogAccelByteLobby, Log, TEXT("fetching lobby error messages DONE! %d lobby messages has been cached"), LobbyErrorMessages.Num());
-	}), FErrorHandler::CreateLambda([&bFetchErrorMessageDone](const int32 code, const FString& message)
+	}), FErrorHandler::CreateLambda([](const int32 code, const FString& message)
 	{
 		UE_LOG(LogAccelByteLobby, Warning, TEXT("Error fetching lobby error messages! code %d, message %s"), code, *message);
-		bFetchErrorMessageDone = true;
 	}));
 }