+++ Source/AccelByteUe4Sdk/Public/Api/AccelByteLobbyApi.h
@@ -1686,7 +1686,7 @@ public:
 	*/
 	void SetTokenGenerator(TSharedPtr<IAccelByteTokenGenerator> TokenGenerator);
 
-	static FString LobbyMessageToJson(FString Message);
+	static FString LobbyMessageToJson(const FString& Message);
 
 	void ClearLobbyErrorMessages();